+++ constants/image.ts
@@ -2,5 +2,6 @@ import logo from "../assets/images/Logo.jpg";
 import logobg from "../assets/images/Logo-removebg.png";
 import path from "../assets/images/path.png"
 import bgimage from "../assets/images/bg.webp"
+import statusImage from "../assets/images/fire.jpg"
 
-export default {logo, logobg, path, bgimage};
+export default {logo, logobg, path, bgimage,statusImage};
